Support non-convex polygons, clipping paths, line form styles, transparency
Uses GLU tesselate function to handle:
    - non-convex polygons
    - self-intersecting polygons
    - fill rules
    - clipping

Converts lines to polygons to support line forms:
    - dashed lines
    - LineCap Attribute
    - LineJoin Attribute

The State Monad is used to allow the final geometry to depend on the
Style.  The state is therefore cleared after each primitive.  Because
the polygons representing lines also need to be clipped, nearly all
the style information needs to be available in the State.  Because a
Path with multiple Trails is handled differently from several
one-Trail Paths, GLU tesselate needs to be called before flattening to
GlPrims.

closes https://github.com/bergey/diagrams-opengl/pull/1